+++ packages/block-library/src/index.js
@@ -148,6 +148,7 @@ export const __experimentalGetCoreBlocks = () => [
 	mediaText,
 	latestComments,
 	latestPosts,
+	legacyWidget,
 	missing,
 	more,
 	nextpage,
@@ -229,14 +230,11 @@ export const registerCoreBlocks = (
  */
 export const __experimentalRegisterExperimentalCoreBlocks =
 	process.env.GUTENBERG_PHASE === 2
-		? ( { enableLegacyWidgetBlock, enableFSEBlocks } = {} ) => {
+		? ( { enableFSEBlocks } = {} ) => {
 				[
 					navigation,
 					navigationLink,
 					homeLink,
-
-					// Register Legacy Widget block.
-					...( enableLegacyWidgetBlock ? [ legacyWidget ] : [] ),
 
 					// Register Full Site Editing Blocks.
 					...( enableFSEBlocks