+++ packages/react-native-bridge/android/react-native-bridge/src/main/java/org/wordpress/mobile/WPAndroidGlue/GutenbergProps.kt
@@ -37,8 +37,6 @@ data class GutenbergProps @JvmOverloads constructor(
             theme.getSerializable(PROP_GRADIENTS)?.let { putSerializable(PROP_GRADIENTS, it) }
             theme.getSerializable(PROP_STYLES)
                     ?.let { putSerializable(PROP_STYLES, it) }
-            theme.getSerializable(PROP_FEATURES)
-                    ?.let { putSerializable(PROP_FEATURES, it) }
         }
     }
 
@@ -72,7 +70,6 @@ data class GutenbergProps @JvmOverloads constructor(
         private const val PROP_COLORS = "colors"
         private const val PROP_GRADIENTS = "gradients"
         private const val PROP_STYLES = "rawStyles"
-        private const val PROP_FEATURES = "rawFeatures"
 
         const val PROP_CAPABILITIES = "capabilities"
         const val PROP_CAPABILITIES_CONTACT_INFO_BLOCK = "contactInfoBlock"