+++ packages/react-native-bridge/ios/GutenbergBridgeDataSource.swift
@@ -73,7 +73,6 @@ public extension GutenbergBridgeDataSource {
 public protocol GutenbergEditorSettings {
     var isFSETheme: Bool { get }
     var rawStyles: String? { get }
-    var rawFeatures: String? { get }
     var colors: [[String: String]]? { get }
     var gradients: [[String: String]]? { get }
 }