+++ phpunit/class-elements-test.php
@@ -4,12 +4,6 @@
  *
  * @package Gutenberg
  */
-
-/**
- * Test WP_Theme_JSON class.
- *
- * @package Gutenberg
- */
 class Gutenberg_Elements_Test extends WP_UnitTestCase {
 	/**
 	 * Given a string containing a class prefixed by "wp-elements-" followed by a unique id,