+++ packages/editor/src/components/error-boundary/index.js
@@ -8,6 +8,11 @@ import { select } from '@wordpress/data';
 import { Warning } from '@wordpress/block-editor';
 import { useCopyToClipboard } from '@wordpress/compose';
 
+/**
+ * Internal dependencies
+ */
+import { store as editorStore } from '../../store';
+
 function CopyButton( { text, children } ) {
 	const ref = useCopyToClipboard( text );
 	return (
@@ -45,7 +50,7 @@ class ErrorBoundary extends Component {
 			// (b) avoids the performance cost associated with unnecessary
 			// content serialization throughout the lifetime of a non-erroring
 			// application.
-			return select( 'core/editor' ).getEditedPostContent();
+			return select( editorStore ).getEditedPostContent();
 		} catch ( error ) {}
 	}